test: working simple and sync
