+++ test/docker-e2e/e2e_state_sync_test.go
@@ -2,21 +2,16 @@ package docker_e2e
 
 import (
 	"context"
-	"github.com/celestiaorg/celestia-app/v4/app"
-	"github.com/chatton/interchaintest"
 	"github.com/chatton/interchaintest/chain/cosmos"
-	"github.com/chatton/interchaintest/ibc"
-	maps "github.com/chatton/interchaintest/testutil/maps"
 	"github.com/chatton/interchaintest/testutil/toml"
 	"github.com/chatton/interchaintest/testutil/wait"
-	"github.com/cosmos/cosmos-sdk/types/module/testutil"
 	"github.com/stretchr/testify/require"
-	"go.uber.org/zap/zaptest"
 	"testing"
 	"time"
 )
 
-func TestCelestiaChainStateSync(t *testing.T) {
+func (s *CelestiaTestSuite) TestCelestiaChainStateSync() {
+	t := s.T()
 	if testing.Short() {
 		t.Skip("skipping in short mode")
 	}
@@ -27,47 +22,8 @@ func TestCelestiaChainStateSync(t *testing.T) {
 		stateSyncTimeout           = 5 * time.Minute
 	)
 
-	// Create a new logger for the test
-	logger := zaptest.NewLogger(t)
-
-	// Setup Docker client and network
-	client, network := interchaintest.DockerSetup(t)
-
-	// Define the number of validators
-	numValidators := 4
-	numFullNodes := 0
-
-	enc := testutil.MakeTestEncodingConfig(app.ModuleEncodingRegisters...)
-	// Create a single Celestia chain directly
-	celestia, err := interchaintest.NewChain(logger, t.Name(), client, network, &interchaintest.ChainSpec{
-		Name:          "celestia",
-		ChainName:     "celestia",
-		Version:       "v4.0.0-rc1",
-		NumValidators: &numValidators,
-		NumFullNodes:  &numFullNodes,
-		Config: ibc.Config{
-			ModifyGenesis: func(config ibc.Config, bytes []byte) ([]byte, error) {
-				return maps.SetField(bytes, "consensus.params.version.app", "4")
-			},
-			EncodingConfig:      &enc,
-			AdditionalStartArgs: []string{"--force-no-bbr", "--grpc.enable", "--grpc.address", "0.0.0.0:9090", "--rpc.grpc_laddr=tcp://0.0.0.0:9099"},
-			Type:                "cosmos",
-			ChainID:             "celestia",
-			Images: []ibc.DockerImage{
-				{
-					Repository: "ghcr.io/celestiaorg/celestia-app-multiplexer",
-					Version:    "v4.0.0-rc1",
-					UIDGID:     "10001:10001",
-				},
-			},
-			Bin:           "celestia-appd",
-			Bech32Prefix:  "celestia",
-			Denom:         "utia",
-			GasPrices:     "0.025utia",
-			GasAdjustment: 1.3,
-		},
-	})
-	require.NoError(t, err)
+	celestia, err := s.CreateCelestiaChain("v4.0.0-rc1", "4")
+	s.Require().NoError(err)
 
 	// Start the chain
 	ctx := context.Background()
@@ -90,12 +46,7 @@ func TestCelestiaChainStateSync(t *testing.T) {
 	cosmosChain, ok := celestia.(*cosmos.Chain)
 	require.True(t, ok, "expected celestia to be a cosmos.Chain")
 
-	// Verify we have the expected number of validators
-	require.Equal(t, numValidators, len(cosmosChain.Validators))
-
-	t.Logf("Successfully started Celestia chain with %d validators", numValidators)
-
-	createTxSim(t, err, ctx, client, network, logger, cosmosChain)
+	s.CreateTxSim(ctx, "v4.0.0-rc1", cosmosChain)
 
 	nodeClient := cosmosChain.Nodes()[0].Client
